backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI and ML
openai==1.12.0
google-generativeai==0.4.0
groq>=0.5.0
tenacity==8.2.3

# Data Validation
pydantic==2.6.0
pydantic-settings==2.1.0

# Text-to-Speech
edge-tts==6.1.9
gTTS>=2.5.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.26.0
Pillow>=10.4.0

# CORS
python-jose==3.3.0
